requirements.txt
tensorflow
numpy
emnist
opencv-python
Pillow